changed typing

+++ OptiScan/rotation_optimisation.py
@@ -141,7 +141,7 @@ def get_optimal_rotation(image, _from=-0.1, _to=0.1, initial_space=0.1, final_sp
 
 def rotate_with_optimal_rotation(image, _from=-0.1, _to=0.1, initial_space=0.05, final_space=0.01, saphyr=False):
     print(True, _from, _to, initial_space, final_space)
-    angle = max(get_optimal_rotation(image.astype(int), _from=_from, _to=_to,
+    angle = max(get_optimal_rotation(image, _from=_from, _to=_to,
                                      initial_space=initial_space, final_space=final_space, saphyr=saphyr))[1]
     return rotate(image, angle), angle
